requirements.txt
Flask==2.0.2
pymongo==3.11.4
Werkzeug==2.0.2
dnspython==2.3.0
python-dotenv==1.0.1
Flask-Mail==0.9.1